refactor: rename peapods parser folder
