+++ daemon/parser.py
@@ -5,7 +5,7 @@ import sys
 from jina import __resources_path__, __default_host__
 from jina.parsers.base import set_base_parser
 from jina.parsers.helper import add_arg_group, _SHOW_ALL_ARGS
-from jina.parsers.peapods.base import mixin_base_ppr_parser
+from jina.parsers.orchestrate.base import mixin_base_ppr_parser
 from daemon.models.enums import PartialDaemonModes
