+++ jina/parsers/__init__.py
@@ -1,6 +1,6 @@
 from jina.parsers.client import mixin_comm_protocol_parser
 from jina.parsers.helper import _SHOW_ALL_ARGS
-from jina.parsers.peapods.runtimes.head import mixin_head_parser
+from jina.parsers.orchestrate.runtimes.head import mixin_head_parser
 
 
 def set_pod_parser(parser=None):
@@ -14,12 +14,14 @@ def set_pod_parser(parser=None):
 
         parser = set_base_parser()
 
-    from jina.parsers.peapods.base import mixin_base_ppr_parser
-    from jina.parsers.peapods.runtimes.worker import mixin_worker_runtime_parser
-    from jina.parsers.peapods.runtimes.container import mixin_container_runtime_parser
-    from jina.parsers.peapods.runtimes.remote import mixin_remote_runtime_parser
-    from jina.parsers.peapods.pod import mixin_pod_parser
-    from jina.parsers.peapods.runtimes.distributed import (
+    from jina.parsers.orchestrate.base import mixin_base_ppr_parser
+    from jina.parsers.orchestrate.runtimes.worker import mixin_worker_runtime_parser
+    from jina.parsers.orchestrate.runtimes.container import (
+        mixin_container_runtime_parser,
+    )
+    from jina.parsers.orchestrate.runtimes.remote import mixin_remote_runtime_parser
+    from jina.parsers.orchestrate.pod import mixin_pod_parser
+    from jina.parsers.orchestrate.runtimes.distributed import (
         mixin_distributed_feature_parser,
     )
     from jina.parsers.hubble.pull import mixin_hub_pull_options_parser
@@ -49,7 +51,7 @@ def set_deployment_parser(parser=None):
 
     set_pod_parser(parser)
 
-    from jina.parsers.peapods.deployment import mixin_base_deployment_parser
+    from jina.parsers.orchestrate.deployment import mixin_base_deployment_parser
 
     mixin_base_deployment_parser(parser)
 
@@ -67,16 +69,16 @@ def set_gateway_parser(parser=None):
 
         parser = set_base_parser()
 
-    from jina.parsers.peapods.base import mixin_base_ppr_parser
-    from jina.parsers.peapods.runtimes.worker import mixin_worker_runtime_parser
-    from jina.parsers.peapods.runtimes.remote import (
+    from jina.parsers.orchestrate.base import mixin_base_ppr_parser
+    from jina.parsers.orchestrate.runtimes.worker import mixin_worker_runtime_parser
+    from jina.parsers.orchestrate.runtimes.remote import (
         mixin_gateway_parser,
         mixin_prefetch_parser,
         mixin_http_gateway_parser,
         mixin_compressor_parser,
     )
-    from jina.parsers.peapods.deployment import mixin_base_deployment_parser
-    from jina.parsers.peapods.pod import mixin_pod_parser
+    from jina.parsers.orchestrate.deployment import mixin_base_deployment_parser
+    from jina.parsers.orchestrate.pod import mixin_pod_parser
 
     mixin_base_ppr_parser(parser)
     mixin_worker_runtime_parser(parser)
@@ -110,7 +112,7 @@ def set_client_cli_parser(parser=None):
 
         parser = set_base_parser()
 
-    from jina.parsers.peapods.runtimes.remote import mixin_client_gateway_parser
+    from jina.parsers.orchestrate.runtimes.remote import mixin_client_gateway_parser
     from jina.parsers.client import (
         mixin_client_features_parser,
         mixin_comm_protocol_parser,
