+++ jina/parsers/orchestrate/runtimes/head.py
@@ -2,7 +2,7 @@ from jina.parsers.helper import add_arg_group
 
 
 def mixin_head_parser(parser):
-    """Mixing in arguments required by head peas and runtimes into the given parser.
+    """Mixing in arguments required by head pods and runtimes into the given parser.
     :param parser: the parser instance to which we add arguments
     """
