+++ jina/proto/__init__.py
@@ -1,4 +1,4 @@
 """
-The :mod:`jina.proto` defines the protobuf used in jina. It is the core message protocol used in communicating between :class:`jina.peapods.base.BaseDeployment`. It also defines the interface of a gRPC service.
+The :mod:`jina.proto` defines the protobuf used in jina. It is the core message protocol used in communicating between :class:`jina.orchestrate.deployments.BaseDeployment`. It also defines the interface of a gRPC service.
 
 """
