+++ tests/unit/parsers/peapods/runtimes/test_remote_parser.py
@@ -1,5 +1,5 @@
 import argparse
-from jina.parsers.peapods.runtimes.remote import mixin_http_gateway_parser
+from jina.parsers.orchestrate.runtimes.remote import mixin_http_gateway_parser
 
 
 def test_runtime_parser():
