+++ tests/unit/parsers/peapods/runtimes/test_runtime_parser.py
@@ -1,5 +1,5 @@
 import argparse
-from jina.parsers.peapods.runtimes.container import mixin_container_runtime_parser
+from jina.parsers.orchestrate.runtimes.container import mixin_container_runtime_parser
 
 
 def test_runtime_parser():
